+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.9 - 2025-12-10
+- Story management improvements:
+    - Adding `publish` parameter in `create()` method for publishing the story immediatly
+    - Adding `releaseId` parameter in the `create()` method for creating the story in a specific release
+    - Adding getter and setter method for parent folder in Story. `setFolderId()` and `folderId()`
+
+
 ## 1.0.8 - 2025-12-08
 - Space class improvement, added some helper methos like `isOwnedByUser()`, `domain()`, `isDemo()`, `fistToken()`, `removeDemoMode()`
 - Added Space methods for handling environments/Preview URLs